RBAC addeed

+++ routes/web.php
@@ -15,11 +15,12 @@ Route::get('/dashboard', function () {
     return view('dashboard');
 })->middleware(['auth', 'verified'])->name('dashboard');
 
-Route::middleware('auth')->group(function () {
+Route::middleware('auth','admin')->group(function () {
     Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
     Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
     Route::get('/dashboard', [DashboardController::class, 'index'])->name('dashboard');
+    Route::resource('departments', DepartmentController::class);
 });
 
 require __DIR__.'/auth.php';
@@ -33,8 +34,8 @@ Route::post('/students', [StudentController::class, 'store'])->name('students.st
 Route::get('/students/{student}', [StudentController::class, 'show'])->name('students.show');
 Route::get('/students/{student}/edit', [StudentController::class, 'edit'])->name('students.edit');
 Route::put('/students/{student}', [StudentController::class, 'update'])->name('students.update');
-Route::delete('/students/{student}', [StudentController::class, 'destroy'])->name('students.destroy');
+Route::delete('/students/{student}', [StudentController::class, 'destroy'])->name('students.destroy')->middleware('admin');;
 
-Route::resource('departments', DepartmentController::class);
+// Route::resource('departments', DepartmentController::class);
 
 Route::post('/check-email', [StudentController::class, 'checkEmail'])->name('students.checkEmail');
